Add forecast & shared module dataclasses :
