+++ backend/fun_forecast_backend/weather/dataclasses.py
@@ -1,20 +1,6 @@
 from typing import Bool, List
 from enum import Enum
-import datetime
 from dataclasses import dataclass
-
-
-class WeatherCondition(Enum):
-    CLEAR = 0
-    MOSTLY_CLEAR = 1
-    PARTLY_CLEAR = 2
-    PARTLY_CLOUDY = 3
-    MOSTLY_CLOUDY = 4
-    CLOUDY = 5
-    FOG = 6
-    RAIN = 7
-    STORM = 8
-    UNKNOWN = 9
 
 
 @dataclass
@@ -71,17 +57,3 @@ class WeatherDataProperties:
 @dataclass
 class WeatherData:
     properties: WeatherDataProperties
-
-
-@dataclass
-class HourData:
-    unit: str = None
-    timezone: str
-    location: str
-    start: datetime.datetime
-    end: datetime.datetime
-    daytime: bool
-    temp: int
-    wind: int
-    weather: str
-    condition: WeatherCondition
